Added Requires annotation and thorough search mode
We can now use "@Require" to dynamically load other modules without the need to inject them as a field first (having fields just for that purpose would be ugly, right?).

This also introduces the "thorough search" mode which was an internal first solution to the problem that, when running the application from inside an IDE, some modules might not work as expected because their requirements weren't loaded beforehand. The new mode can be enabled either by code or commandline (using the "-Dmodular.thoroughSearch" property). However, this will increase the loading time drasticly which is why the new annotation should be used instead.