Fixed .class files not getting recognized properly
Since we never set the root for the files to begin with their "internal package name" was always empty, and thus they never were discovered.

We now set the given URI as the initial root and and it to the realm whenever we found a class file so its package name is properly discovered.

A workaround would have been using something like a .class-reader to get the package name from the file directly, but would only slow down the loading process.