Reworked the injector API to be way more intuitive
The new injection API has been reworked (and simplified) with the concept of a builder in mind in order to be more flexible and intuitive.

Similar to google's Guice it's now as simple as saying "inject(Thing.class).with(instance)". No more supportsor entries are needed and/or required.

Please check the javadocs for more information.